Added Sound. NEEDS AUDIO CHANGE OUT OF UPDATE

+++ PA10/include/object.h
@@ -78,7 +78,7 @@ class Object
     btCollisionShape *colliderShape;
     btTriangleMesh *objTriMesh;
     btRigidBody* rigidBody;
-    int m_mass;
+    float m_mass;
     
     // If the object has a child, this is the matrix it should read as its center
     glm::mat4 modelForChild;
